v1.0.0.100
+ Added gitignore.
+ Added read me.
* Cleaned up unused files.
* Fixed up some addresses.
* Refactored some code.
! At this state, the game can load NFS Underground platform with its exclusive features which was the basic idea. It can also switch between vanilla and NFS Underground platform.
