+++ NFSMW_FELoader/NFSMWFELoader.ini
@@ -1,0 +1,85 @@
+[GENERAL]
+LoadMapInFE = 0
+LoadCustomPlatform = 1
+
+[CUSTOM_PLATFORM]
+Path = FRONTEND\PLATFORMS\UG1.BIN
+X = 0
+Y = 0
+Z = -0.25
+ScrollItems = 25
+ScrollSpeedMin = 1
+ScrollSpeedMax = 10
+
+[PLATFORMCRIB]
+Path = FRONTEND\PLATFORMS\PLATFORMCRIB.BIN
+
+[CAREER_SAFEHOUSE]
+Path = FRONTEND\PLATFORMS\CAREER_SAFEHOUSE.BIN
+
+[CUSTOMIZATION_SHOP]
+Path = FRONTEND\PLATFORMS\CUSTOMIZATION_SHOP.BIN
+
+[CUSTOMIZATION_SHOP_BACKROOM]
+Path = FRONTEND\PLATFORMS\CUSTOMIZATION_SHOP_BACKROOM.BIN
+
+[CAR_LOT]
+Path = FRONTEND\PLATFORMS\CAR_LOT.BIN
+
+// In game Front End car positions
+[POS_MAIN]
+Preset = 0
+
+[POS_PLATFORMCRIB]
+Preset = 0
+
+[POS_CAREER_SAFEHOUSE]
+Preset = 2
+
+[POS_CUSTOMIZATION_SHOP]
+Preset = 3
+
+[POS_CUSTOMIZATION_SHOP_BACKROOM]
+Preset = 4
+
+[POS_CAR_LOT]
+Preset = 5
+
+// Position presets
+[PRESET_0] // Custom
+X = 0
+Y = 0
+Z = -0.3
+CarZ = 0
+CarReflection = 0 // Enabled by default
+CustomPlatform = 1 // Indicates that custom platfrom should be rendered for this preset
+
+[PRESET_1] // PlatformCrib (Main Menu)
+X = 0
+Y = 0
+Z = 0
+CarZ = 304.96979
+
+[PRESET_2] // Career Safehouse
+X = 0
+Y = 0
+Z = 0
+CarZ = 304.96979
+
+[PRESET_3] // Customization Shop
+X = 0
+Y = 0
+Z = 0
+CarZ = 304.96979
+
+[PRESET_4] // Customization Shop Backroom
+X = 0
+Y = 0
+Z = 0
+CarZ = 304.96979
+
+[PRESET_5] // Car Lot
+X = 0
+Y = 0.075
+Z = 0
+CarZ = 340
