+++ NFSMW_FELoader/NFSMW_FELoader.vcxproj.user
@@ -1,4 +1,13 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup />
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
+    <LocalDebuggerCommand>G:\Oyunlar\Need for Speed Most Wanted Dirty\speed.exe</LocalDebuggerCommand>
+    <LocalDebuggerWorkingDirectory>G:\Oyunlar\Need for Speed Most Wanted Dirty\</LocalDebuggerWorkingDirectory>
+    <DebuggerFlavor>WindowsLocalDebugger</DebuggerFlavor>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
+    <LocalDebuggerCommand>G:\Oyunlar\Need for Speed Most Wanted Dirty\speed.exe</LocalDebuggerCommand>
+    <LocalDebuggerWorkingDirectory>G:\Oyunlar\Need for Speed Most Wanted Dirty\</LocalDebuggerWorkingDirectory>
+    <DebuggerFlavor>WindowsLocalDebugger</DebuggerFlavor>
+  </PropertyGroup>
 </Project>
