+++ README.md
@@ -1,2 +1,13 @@
-# NFSMW_FELOADER
-WIP... Don't use this, yet...
+# NFSMW FrontEnd Platform Loader
+
+## Fork Note
+This project is a port of @ArchieGoldmill's [NFSC_FELoader](https://github.com/ArchieGoldmill/NFSC_FELoader) to Need for Speed: Most Wanted (2005).  
+
+
+## To-Do
+!!! WIP... Don't use this, yet...  
+[] Fix not working features  
+[] Allow custom platforms per garage type (It currently supports only one)  
+[] Clean up unused stuff  
+[] Switch to mINI  
+[] Test if it works with previously released platform mods.
