--- Release/NFSCFELoader.asi.recipe
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project>
-  <ProjectOutputs>
-    <ProjectOutput>
-      <FullPath>D:\PS2\NFSMW_FELoader-master\NFSMW_FELoader\scripts\NFSCFELoader.asi</FullPath>
-    </ProjectOutput>
-  </ProjectOutputs>
-  <ContentFiles />
-  <SatelliteDlls />
-  <NonRecipeFileRefs />
-</Project>
