--- Release/NFSMW_FELoader.log
@@ -1,10 +1,0 @@
-﻿C:\Program Files\Microsoft Visual Studio\2022\Community\MSBuild\Microsoft\VC\v170\Microsoft.CppBuild.targets(500,5): warning MSB8004: o diretório Output não termina com uma barra à direita. Essa instância de compilação adicionará a barra conforme necessário para permitir a avaliação adequada do Diretório Output.
-  dllmain.cpp
-D:\PS2\NFSMW_FELoader-master\NFSMW_FELoader\dllmain.cpp(319,49): warning C4244: '+=': conversão de 'double' para 'float', possível perda de dados
-D:\PS2\NFSMW_FELoader-master\NFSMW_FELoader\dllmain.cpp(408,59): warning C4244: 'argumento': conversão de 'double' para 'FLOAT', possível perda de dados
-D:\PS2\NFSMW_FELoader-master\NFSMW_FELoader\dllmain.cpp(474,38): warning C4244: '=': conversão de 'float' para 'int', possível perda de dados
-  Gerando código
-  Previous IPDB not found, fall back to full compilation.
-  All 888 functions were compiled because no usable IPDB/IOBJ from previous compilation was found.
-  Finalizada a geração de código
-  NFSMW_FELoader.vcxproj -> D:\PS2\NFSMW_FELoader-master\NFSMW_FELoader\scripts\NFSMWFELoader.asi
