--- Release/NFSMW_FELoader.tlog/NFSMW_FELoader.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native32Bit:VCToolsVersion=14.32.31326:TargetPlatformVersion=10.0.22000.0:
-Release|Win32|D:\PS2\NFSMW_FELoader-master\|
